preliminary action to deck and board class
