+++ src/main/java/models/Board.java
@@ -1,0 +1,64 @@
+package models;
+
+import java.util.ArrayList;
+
+public class Board {
+
+    public java.util.List<java.util.List<Card>> cols = new ArrayList<>();
+
+    public Board()
+    {
+        cols.add(new ArrayList<Card>());
+        cols.add(new ArrayList<Card>());
+        cols.add(new ArrayList<Card>());
+        cols.add(new ArrayList<Card>());
+    }
+
+    public void remove(int columnNumber) {
+        if(columnHasCards(columnNumber)) {
+            Card c = getTopCard(columnNumber);
+            boolean removeCard = false;
+            for (int i = 0; i < 4; i++) {
+                if (i != columnNumber) {
+                    if (columnHasCards(i)) {
+                        Card compare = getTopCard(i);
+                        if (compare.getSuit() == c.getSuit()) {
+                            if (compare.getValue() > c.getValue()) {
+                                removeCard = true;
+                            }
+                        }
+                    }
+                }
+            }
+            if (removeCard) {
+                this.cols.get(columnNumber).remove(this.cols.get(columnNumber).size() - 1);
+            }
+        }
+    }
+
+    private boolean columnHasCards(int columnNumber) {
+        if(this.cols.get(columnNumber).size()>0){
+            return true;
+        }
+        return false;
+    }
+
+    private Card getTopCard(int columnNumber) {
+        return this.cols.get(columnNumber).get(this.cols.get(columnNumber).size()-1);
+    }
+
+
+    public void move(int columnFrom, int columnTo) {
+        Card cardToMove = getTopCard(columnFrom);
+        this.removeCardFromCol(columnFrom);
+        this.addCardToCol(columnTo,cardToMove);
+    }
+
+    private void addCardToCol(int columnTo, Card cardToMove) {
+        cols.get(columnTo).add(cardToMove);
+    }
+
+    private void removeCardFromCol(int colFrom) {
+        this.cols.get(colFrom).remove(this.cols.get(colFrom).size()-1);
+    }
+}
