+++ src/main/java/models/Game.java
@@ -14,14 +14,11 @@ public class Game {
 
     public java.util.List<Card> deck = new ArrayList<>();
 
-    public java.util.List<java.util.List<Card>> cols = new ArrayList<>();
-
+    public Board gameBoard; //object to hold the columns which hold cards
 
     public Game(){
-        cols.add(new ArrayList<Card>());
-        cols.add(new ArrayList<Card>());
-        cols.add(new ArrayList<Card>());
-        cols.add(new ArrayList<Card>());
+
+        gameBoard = new Board(); // initialize a new board object for the game
     }
 
     public void buildDeck() {
@@ -38,15 +35,18 @@ public class Game {
         Collections.shuffle(deck, new Random(seed));
     }
 
-    public void dealFour() {
+    public void dealFour(Board gameBoard) {
         for(int i = 0; i < 4; i++){
-            cols.get(i).add(deck.get(deck.size()-1));
+            gameBoard.cols.get(i).add(deck.get(deck.size()-1));
             deck.remove(deck.size()-1);
+            System.out.println("HERRO");
+            //System.out.println(gameBoard);
+            //System.out.println(gameBoard.cols);
         }
     }
 
     //customDeal to setup game for testing purposes
-    public void customDeal(int c1, int c2, int c3, int c4) {
+    /*public void customDeal(int c1, int c2, int c3, int c4) {
         cols.get(0).add(deck.get(c1));
         deck.remove(c1);
         cols.get(1).add(deck.get(c2));
@@ -55,53 +55,5 @@ public class Game {
         deck.remove(c3);
         cols.get(3).add(deck.get(c4));
         deck.remove(c4);
-    }
-
-    public void remove(int columnNumber) {
-        if(columnHasCards(columnNumber)) {
-            Card c = getTopCard(columnNumber);
-            boolean removeCard = false;
-            for (int i = 0; i < 4; i++) {
-                if (i != columnNumber) {
-                    if (columnHasCards(i)) {
-                        Card compare = getTopCard(i);
-                        if (compare.getSuit() == c.getSuit()) {
-                            if (compare.getValue() > c.getValue()) {
-                                removeCard = true;
-                            }
-                        }
-                    }
-                }
-            }
-            if (removeCard) {
-                this.cols.get(columnNumber).remove(this.cols.get(columnNumber).size() - 1);
-            }
-        }
-    }
-
-    private boolean columnHasCards(int columnNumber) {
-        if(this.cols.get(columnNumber).size()>0){
-            return true;
-        }
-        return false;
-    }
-
-    private Card getTopCard(int columnNumber) {
-        return this.cols.get(columnNumber).get(this.cols.get(columnNumber).size()-1);
-    }
-
-
-    public void move(int columnFrom, int columnTo) {
-        Card cardToMove = getTopCard(columnFrom);
-        this.removeCardFromCol(columnFrom);
-        this.addCardToCol(columnTo,cardToMove);
-    }
-
-    private void addCardToCol(int columnTo, Card cardToMove) {
-        cols.get(columnTo).add(cardToMove);
-    }
-
-    private void removeCardFromCol(int colFrom) {
-        this.cols.get(colFrom).remove(this.cols.get(colFrom).size()-1);
-    }
+    }*/
 }
